requirements.txt
# LLM / Embedding
ollama

# Vector DB
chromadb

# PDF manuals
pypdf

# CSV / Excel trouble cases
pandas
openpyxl

# Config management
python-dotenv

# Simple Web UI
streamlit

# HTTP client
requests